Move error classes to seperate files for better 3rd party PSR-0 compability
