+++ src/Balanced/Errors/AccountIsAlreadyAMerchant.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class AccountIsAlreadyAMerchant extends Error
+{
+    public static $codes = array('account-already-merchant');
+}
