+++ src/Balanced/Errors/AddressVerificationFailed.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class AddressVerificationFailed extends Error
+{
+    public static $codes = array('address-verification-failed');
+}
