+++ src/Balanced/Errors/CannotAssociateCard.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class CannotAssociateCard extends Error
+{
+    public static $codes = array('cannot-associate-card');
+}
