+++ src/Balanced/Errors/CannotAssociateMerchantWithAccount.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class CannotAssociateMerchantWithAccount extends Error
+{
+    public static $codes = array('cannot-associate-merchant-with-account');
+}
