+++ src/Balanced/Errors/CannotCredit.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class CannotCredit extends Error
+{
+    public static $codes = array('funding-destination-not-creditable');
+}
