+++ src/Balanced/Errors/CannotHold.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class CannotHold extends Error
+{
+    public static $codes = array('funding-source-not-hold');
+}
