+++ src/Balanced/Errors/DuplicateAccountEmailAddress.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class DuplicateAccountEmailAddress extends Error
+{
+    public static $codes = array('duplicate-email-address');
+}
