+++ src/Balanced/Errors/HoldExpired.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class HoldExpired extends Error
+{
+    public static $codes = array('authorization-expired');
+}
