+++ src/Balanced/Errors/InsufficientFunds.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class InsufficientFunds extends Error
+{
+    public static $codes = array('insufficient-funds');
+}
