+++ src/Balanced/Errors/InvalidAmount.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class InvalidAmount extends Error
+{
+    public static $codes = array('invalid-amount');
+}
