+++ src/Balanced/Errors/InvalidBankAccountNumber.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class InvalidBankAccountNumber extends Error
+{
+    public static $codes = array('invalid-bank-account-number');
+}
