+++ src/Balanced/Errors/InvalidRoutingNumber.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class InvalidRoutingNumber extends Error
+{
+    public static $codes = array('invalid-routing-number');
+}
