+++ src/Balanced/Errors/MarketplaceAlreadyCreated.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class MarketplaceAlreadyCreated extends Error
+{
+    public static $codes = array('marketplace-already-created');
+}
