+++ src/Balanced/Errors/NoFundingDestination.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class NoFundingDestination extends Error
+{
+    public static $codes = array('no-funding-destination');
+}
