+++ src/Balanced/Errors/NoFundingSource.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Balanced\Errors;
+
+use Balanced\Errors\Error;
+
+class NoFundingSource extends Error
+{
+    public static $codes = array('no-funding-source');
+}
